Add deprecation message to service:* help commands
It would have been great to add the full deprecation message
to the top of the service:* command help output, but oclif
will only allow one line to be shown in the description section
of the help. As soon as it encounters a newline, it moves all
subsequent content down into a DESCRIPTION section. To help with
this, this commit prefixes each service:* description with
[DEPRECATED] to make sure the deprecation notice is one of the
first things people see. Then the full deprecation message is
shown in the DESCRIPTION section.