apollo-tools: remove dependency on apollo-env (#2339)
As far as I can tell, the only use `@apollographql/apollo-tools` makes
of `apollo-env` is the ability to use `Array.flat` in pre-12 Node. Now,
we probably should just make this whole repo require Node 12 like we are
doing for Apollo Server 3. But a simpler way to break this dependency is
to just implement a single-level Array.flat ourselves. (The calls to
Array.flat do not take advantage of its ability to flatten more than one
level of nested array.)

This will allow us to stop installing polyfills in Apollo Server 3 (see
apollographql/apollo-server#2634).