Handle missing upsertImplementingServiceAndTriggerComposition field
After re-running codegen based on the latest Apollo Studio API schema,
it's now possible that the `UploadAndComposePartialSchema_service`
interface returns `null` for
`upsertImplementingServiceAndTriggerComposition`:

```
export interface UploadAndComposePartialSchema_service {
  __typename: "ServiceMutation";
  upsertImplementingServiceAndTriggerComposition: UploadAndComposePartialSchema_service_upsertImplementingServiceAndTriggerComposition | null;
}
```

The language server in this project depends on
`upsertImplementingServiceAndTriggerComposition` being available,
which is why the types currently checked into `master` are causing
build errors.

This PR adds an extra guard to ensure the
`upsertImplementingServiceAndTriggerComposition` field is not null
before proceeding, and throws an exception if not (since it is
required). I've looked through the Studio API code and noticed
that we're now doing similar safety checks on this field, so this
seems like the best solution for now (and shouldn't break anything
consumer wise, as the possibility of `null` coming back has been
in place for quite some time).