Remove `moment` as a dependency for `apollo-language-server` (#2595)
There is really no strong need for `moment` to a dependency of `apollo-language-server` when it is used just to calculate durations in milliseconds. A simple constant will suffice instead.

Also added some tests to prove that this change works but is by no means exhaustive.

Related: https://github.com/apollographql/apollo-tooling/issues/2323.